add 'voted' filter, move code to separate file
