+++ src/app/redux/FetchDataSaga.js
@@ -8,11 +8,6 @@ import {
 } from 'redux-saga/effects';
 import { api } from '@blocktradesdev/steem-js';
 import { loadFollows } from 'app/redux/FollowSaga';
-import {
-    getContent,
-    listProposals,
-    listVoterProposals,
-} from 'app/redux/SagaShared';
 import * as globalActions from './GlobalReducer';
 import * as appActions from './AppReducer';
 import constants from './constants';
@@ -20,8 +15,6 @@ import { fromJS, Map, Set } from 'immutable';
 import { getStateAsync } from 'app/utils/steemApi';
 
 const REQUEST_DATA = 'fetchDataSaga/REQUEST_DATA';
-const LIST_PROPOSALS = 'fetchDataSaga/LIST_PROPOSALS';
-const LIST_VOTER_PROPOSALS = 'fetchDataSaga/LIST_VOTER_PROPOSALS';
 const GET_CONTENT = 'fetchDataSaga/GET_CONTENT';
 const FETCH_STATE = 'fetchDataSaga/FETCH_STATE';
 
@@ -29,17 +22,7 @@ export const fetchDataWatches = [
     takeLatest('@@router/LOCATION_CHANGE', fetchState),
     takeLatest(FETCH_STATE, fetchState),
     takeEvery('global/FETCH_JSON', fetchJson),
-    takeEvery(LIST_PROPOSALS, listProposalsCaller),
-    takeEvery(LIST_VOTER_PROPOSALS, listVoterProposalsCaller),
 ];
-
-export function* listProposalsCaller(action) {
-    yield listProposals(action.payload);
-}
-
-export function* listVoterProposalsCaller(action) {
-    yield listVoterProposals(action.payload);
-}
 
 let is_initial_state = true;
 export function* fetchState(location_change_action) {
@@ -172,15 +155,6 @@ export const actions = {
 
     fetchState: payload => ({
         type: FETCH_STATE,
-        payload,
-    }),
-    listProposals: payload => ({
-        type: LIST_PROPOSALS,
-        payload,
-    }),
-
-    listVoterProposals: payload => ({
-        type: LIST_VOTER_PROPOSALS,
         payload,
     }),
 };
