+++ src/app/redux/SagaShared.js
@@ -68,90 +68,6 @@ function* showTransactionErrorNotification() {
     }
 }
 
-export function* listProposals({
-    start,
-    order_by,
-    order_direction,
-    limit,
-    status,
-    last_id,
-    resolve,
-    reject,
-}) {
-    let proposals;
-    while (!proposals) {
-        proposals = yield call(
-            [api, api.listProposalsAsync],
-            start,
-            order_by,
-            order_direction,
-            limit,
-            status,
-            last_id
-        );
-    }
-
-    yield put(globalActions.receiveListProposals({ proposals }));
-    if (resolve && proposals) {
-        resolve(proposals);
-    } else if (reject && !proposals) {
-        reject();
-    }
-}
-
-export function* listVoterProposals({
-    start,
-    order_by,
-    order_direction,
-    limit,
-    status,
-    resolve,
-    reject,
-}) {
-    let voterProposals = { [start]: [] };
-    let last_id = null;
-    let isLast = false;
-
-    while (!isLast) {
-        const data = yield call(
-            [api, api.listVoterProposalsAsync],
-            start,
-            order_by,
-            order_direction,
-            limit,
-            status,
-            last_id
-        );
-
-        if (data) {
-            if (!data.hasOwnProperty(start)) {
-                isLast = true;
-            } else {
-                let proposals = [];
-
-                if (data[start].length < limit) {
-                    proposals = [...voterProposals[start], ...data[start]];
-                    isLast = true;
-                } else {
-                    const nextProposals = [...data[start]];
-                    last_id = nextProposals[nextProposals.length - 1]['id'];
-                    nextProposals.splice(-1, 1);
-                    proposals = [...voterProposals[start], ...nextProposals];
-                }
-
-                voterProposals = { [start]: proposals };
-            }
-        }
-    }
-
-    yield put(globalActions.receiveListVoterProposals({ voterProposals }));
-    if (resolve && voterProposals[start].length > 0) {
-        resolve(voterProposals);
-    } else if (reject && !voterProposals) {
-        reject();
-    }
-}
-
 /**
  * Save this user's preferences, either directly from the submitted payload or from whatever's saved in the store currently.
  *
